==> 3 - Movie Feature Engineering.ipynb <==
import pandas as pd 

df = pd.read_csv("movies_final.csv")

df.head()

# create ratio between revenue and budget
df["Revenue-Budget Ratio"] = df["revenue"]/df["budget"]

df.head()

# create ratio between revenue and runtime
df["Revenue-Runtime Ratio"] = df["revenue"]/df["runtime"]

df.head()

# import nltk vader library
from nltk.sentiment.vader import SentimentIntensityAnalyzer

# set title as corpus
corpus = df['Title']

# initiate an analyzer
sia = SentimentIntensityAnalyzer()

desc_compound = []
desc_neu = []
desc_neg = []
desc_pos = []

# iterate through each sentence in corpus
for sentence in corpus:
    
    #print(sentence)
    
    # analyze the sentiment. ss is a dictionary
    ss = sia.polarity_scores(sentence)
    
    # output each sentiment score (neg, neu, pos, compound) in ss
    desc_compound.append(ss['compound'])
    desc_neu.append(ss['neu'])
    desc_neg.append(ss['neg'])
    desc_pos.append(ss['pos'])

# add sentiment scores into data frame
df = df.assign(compound = desc_compound)
df = df.assign(neutral = desc_neu)
df = df.assign(negative = desc_neg)
df = df.assign(positive = desc_pos)

# remove '/100' and convert to numeric
df['Rotten Tomatoes'] = df['Rotten Tomatoes'].str.replace('/100', '')
df['Rotten Tomatoes'] = pd.to_numeric(df['Rotten Tomatoes'])

df.head(10)

# Saving changes to CSV
df.to_csv("movies_final.csv", index=False)
