==> 4 - Overview Sentiment Analysis.ipynb <==
# Reading in dataset
df = pd.read_csv("movies_final.csv")

# Renaming sentiment analysis column names for tagline
df = df.rename(columns={'compound': 'tagline_compound', 'neutral': 'tagline_neutral', 'negative': 'tagline_negative', 'positive': 'tagline_positive'})
df.columns # checking columns 

### Sentiment Analysis on Overview

# Initializing the column for sentiment analysis
corpus = df['overview']

# import nltk vader library
from nltk.sentiment.vader import SentimentIntensityAnalyzer
    
# initiate an analyzer
sia = SentimentIntensityAnalyzer()

description_pos = []
description_neg = []
description_neu = []
description_compound = []

# iterate through each sentence in corpus
for sentence in corpus:
    
    #print(sentence)
    
    # analyze the sentiment. ss is a dictionary
    ss = sia.polarity_scores(sentence)
    
    # output each sentiment score (neg, neu, pos, compound) in ss
    #print(ss['pos'])
    description_pos.append(ss['pos'])
    description_neg.append(ss['neg'])
    description_neu.append(ss['neu'])
    description_compound.append(ss['compound'])
    
    # print an empty line as seperator
    #print('\n')

# Creating new columns in df for each score
df = df.assign(overview_positive = description_pos)
df = df.assign(overview_negative = description_neg)
df = df.assign(overview_neutral = description_neu)
df = df.assign(overview_compound = description_compound)
df.columns

# Showing preview of update df
df.head()

# Saving changes to CSV
df.to_csv("movies_final.csv", index=False)
